Refactor/global styles (#43)
* chore: update global styling, removed custom elements, added mobile friendly css for the header, added node-sass

* chore: fixed formatting

* Update frontend/src/components/Header/Header.scss
