+++ frontend/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@code4ro/taskforce-fe-components": "^1.0.0",
     "bulma": "^0.8.0",
+    "node-sass": "^4.13.1",
     "react": "^16.13.0",
     "react-bulma-components": "^3.2.0",
     "react-dom": "^16.13.0",
