+++ frontend/src/components/Header/Header.scss
@@ -1,0 +1,77 @@
+.top-header {
+  display: grid;
+  grid-template-columns: 100%;
+  border-bottom: 1px solid #e0e0e0;
+  align-items: center;
+  justify-content: center;
+
+  .app-logo {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+  }
+
+  nav {
+    display: flex;
+    justify-content: center;
+  }
+
+  a {
+    font-size: 14px;
+    line-height: 21px;
+    display: flex;
+    align-items: center;
+    text-align: center;
+    color: #000000;
+    text-decoration: none;
+  }
+}
+
+.app-credits {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 10px 0;
+  border-bottom: 1px solid #e0e0e0;
+
+  span {
+    padding: 0 10px;
+  }
+}
+
+.bottom-header {
+  display: flex;
+  justify-content: center;
+}
+
+@media (min-width:480px) { 
+  .top-header {
+    display: grid;
+    grid-template-rows: 100%;
+    grid-template-columns: 212px calc(100% - 212px);
+    border-bottom: 1px solid #e0e0e0;
+    align-items: center;
+    nav {
+      display: flex;
+      justify-content: flex-end;
+      padding: 10px;
+    }
+  }
+  .app-credits {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    margin-right: 40px;
+    padding: 10px 0;
+    border-bottom: 1px solid #e0e0e0;
+
+    span {
+      padding: 0 10px;
+    }
+  }
+  .bottom-header {
+    display: flex;
+    justify-content: flex-end;
+  }
+}
+
