+++ frontend/src/index.scss
@@ -1,3 +1,6 @@
+* {
+  font-family: 'Titillium Web', sans-serif;
+}
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
